fix: config if entry type unsoundness.
