+++ test/grammar/types/var_type_annotation/type_fail_13/main.k
@@ -1,0 +1,4 @@
+config: {"A"|"B": int} = {
+    if True:
+        A = "2"
+}
