+++ test/grammar/types/var_type_annotation/type_fail_13/stderr.golden.py
@@ -1,0 +1,22 @@
+
+import sys
+import kclvm.kcl.error as kcl_error
+import os
+
+cwd = os.path.dirname(os.path.realpath(__file__))
+
+kcl_error.print_kcl_error_message(
+    kcl_error.get_exception(
+        err_type=kcl_error.ErrType.TypeError_Compile_TYPE,
+        file_msgs=[
+            kcl_error.ErrFileMsg(
+                filename=cwd + "/main.k",
+                line_no=1,
+                col_no=1,
+                arg_msg="got {str:str(2)}"
+            )
+        ],
+        arg_msg="expect {str(A)|str(B):int}, got {str:str(2)}"
+    ),
+    file=sys.stdout
+)
